requirements.txt
discord.py>=2.3.0
python-dotenv>=1.0.0
aiosqlite>=0.19.0
pytest>=8.2.0
pytest-asyncio>=0.23.8
